Last Push

+++ src/main/java/com/restproject/backend/dtos/request/UpdateExerciseRequest.java
@@ -33,6 +33,6 @@ public class UpdateExerciseRequest {
 
     @NotEmpty
     @NotNull
-    @ListTypeConstraint(type = Integer.class)
+    @ListTypeConstraint(type = Long.class)
     Collection<Long> muscleIds;
 }
